Added a bunch of half baked functionality
Random - randomizes the "spray" selection
Spray - creates a range of possible grains, useless w/o random currently

And also added EFFECTS

Reverb - currently doesn't seem to work so may reimplement
Bitcrusher - works I think??

Will need to rewrite and clean up a lot of this, this was more about
seeing something happen on the screen then doing it right.

+++ synth-slider.js
@@ -33,7 +33,7 @@ export default class SynthSlider extends HTMLElement {
 
 
     const updateConfigEvent = new CustomEvent("update-config", {
-      detail: { name, value: parseInt(value, 10) },
+      detail: { name, value: parseFloat(value, 10) },
       bubbles: true,
       composed: true,
     });
